feat: retire Lurk Proving
