+++ benches/fibonacci.rs
@@ -1,5 +1,6 @@
 use std::{cell::RefCell, rc::Rc, sync::Arc, time::Duration};
 
+use anyhow::anyhow;
 use criterion::{
     black_box, criterion_group, criterion_main, measurement, BatchSize, BenchmarkGroup,
     BenchmarkId, Criterion, SamplingMode,
@@ -8,21 +9,16 @@ use criterion::{
 use pasta_curves::pallas;
 
 use lurk::{
-    circuit::circuit_frame::MultiFrame,
-    eval::{
-        empty_sym_env,
-        lang::{Coproc, Lang},
-    },
+    eval::lang::{Coproc, Lang},
     field::LurkField,
+    lem::{eval::evaluate, multiframe::MultiFrame, pointers::Ptr, store::Store},
     proof::nova::NovaProver,
     proof::Prover,
-    ptr::Ptr,
     public_parameters::{
         instance::{Instance, Kind},
         public_params,
     },
     state::State,
-    store::Store,
 };
 
 mod common;
@@ -35,7 +31,7 @@ fn fib<F: LurkField>(store: &Store<F>, state: Rc<RefCell<State>>, _a: u64) -> Pt
   (fib))
 "#;
 
-    store.read_with_state(state, program).unwrap()
+    store.read(state, program).unwrap()
 }
 
 // The env output in the `fib_frame`th frame of the above, infinite Fibonacci computation will contain a binding of the
@@ -60,30 +56,64 @@ struct ProveParams {
 }
 
 impl ProveParams {
-    fn name(&self) -> String {
-        format!("Fibonacci-rc={}", self.reduction_count)
+    fn name_params(&self) -> (String, String) {
+        let output_type = bench_parameters_env().unwrap_or("stdout".into());
+        match output_type.as_ref() {
+            "pr-comment" => ("fib".into(), format!("num-{}", self.fib_n)),
+            "commit-comment" => (
+                format!("fib-ref={}", env!("VERGEN_GIT_SHA")),
+                format!("num-{}", self.fib_n),
+            ),
+            // TODO: refine "gh-pages",
+            _ => (
+                "fib".into(),
+                format!("num-{}-{}-{}", self.fib_n, self.sha, self.date),
+            ),
+        }
     }
 }
 
-fn fibo_prove<M: measurement::Measurement>(
+fn bench_parameters_env() -> anyhow::Result<String> {
+    std::env::var("LURK_BENCH_OUTPUT")
+        .map_err(|e| anyhow!("Noise threshold env var isn't set: {e}"))
+}
+
+fn rc_env() -> anyhow::Result<Vec<usize>> {
+    std::env::var("LURK_RC")
+        .map_err(|e| anyhow!("Reduction count env var isn't set: {e}"))
+        .and_then(|rc| {
+            let vec: anyhow::Result<Vec<usize>> = rc
+                .split(',')
+                .map(|rc| {
+                    rc.parse::<usize>()
+                        .map_err(|e| anyhow!("Failed to parse RC: {e}"))
+                })
+                .collect();
+            vec
+        })
+}
+
+fn noise_threshold_env() -> anyhow::Result<f64> {
+    std::env::var("LURK_BENCH_NOISE_THRESHOLD")
+        .map_err(|e| anyhow!("Noise threshold env var isn't set: {e}"))
+        .and_then(|nt| {
+            nt.parse::<f64>()
+                .map_err(|e| anyhow!("Failed to parse noise threshold: {e}"))
+        })
+}
+
+fn fibonacci_prove<M: measurement::Measurement>(
     prove_params: ProveParams,
     c: &mut BenchmarkGroup<'_, M>,
     state: &Rc<RefCell<State>>,
 ) {
-    let ProveParams {
-        fib_n,
-        reduction_count,
-        date,
-        sha,
-    } = prove_params;
-
-    let limit = fib_limit(fib_n, reduction_count);
+    let limit = fib_limit(prove_params.fib_n, prove_params.reduction_count);
     let lang_pallas = Lang::<pallas::Scalar, Coproc<pallas::Scalar>>::new();
     let lang_rc = Arc::new(lang_pallas.clone());
 
     // use cached public params
     let instance = Instance::new(
-        reduction_count,
+        prove_params.reduction_count,
         lang_rc.clone(),
         true,
         Kind::NovaPublicParams,
@@ -91,21 +121,18 @@ fn fibo_prove<M: measurement::Measurement>(
     let pp = public_params::<_, _, MultiFrame<'_, _, _>>(&instance).unwrap();
 
     // Track the number of `Lurk frames / sec`
-    let rc = reduction_count as u64;
+    let rc = prove_params.reduction_count as u64;
     c.throughput(criterion::Throughput::Elements(
-        rc * u64::div_ceil((11 + 16 * fib_n) as u64, rc),
+        rc * u64::div_ceil((11 + 16 * prove_params.fib_n) as u64, rc),
     ));
+    let (name, params) = prove_params.name_params();
 
     c.bench_with_input(
-        BenchmarkId::new(
-            prove_params.name(),
-            format!("num-{}/{sha}-{date}", prove_params.fib_n),
-        ),
+        BenchmarkId::new(name, params),
         &prove_params,
         |b, prove_params| {
             let store = Store::default();
 
-            let env = empty_sym_env(&store);
             let ptr = fib::<pasta_curves::Fq>(
                 &store,
                 state.clone(),
@@ -113,9 +140,10 @@ fn fibo_prove<M: measurement::Measurement>(
             );
             let prover = NovaProver::new(prove_params.reduction_count, lang_pallas.clone());
 
-            let frames = &prover
-                .get_evaluation_frames(ptr, env, &store, limit, lang_rc.clone())
-                .unwrap();
+            let frames =
+                &evaluate::<pasta_curves::Fq, Coproc<pasta_curves::Fq>>(None, ptr, &store, limit)
+                    .unwrap()
+                    .0;
 
             b.iter_batched(
                 || (frames, lang_rc.clone()),
@@ -129,25 +157,32 @@ fn fibo_prove<M: measurement::Measurement>(
     );
 }
 
-fn fibonacci_prove(c: &mut Criterion) {
+fn fibonacci_benchmark(c: &mut Criterion) {
+    // Uncomment to record the logs. May negatively impact performance
+    //tracing_subscriber::fmt::init();
     set_bench_config();
     tracing::debug!("{:?}", lurk::config::LURK_CONFIG);
-    let reduction_counts = [100, 600, 700, 800, 900];
+
+    let reduction_counts = rc_env().unwrap_or_else(|_| vec![100]);
     let batch_sizes = [100, 200];
-    let mut group: BenchmarkGroup<'_, _> = c.benchmark_group("Prove");
-    group.sampling_mode(SamplingMode::Flat); // This can take a *while*
-    group.sample_size(10);
+
     let state = State::init_lurk_state().rccell();
 
-    for fib_n in batch_sizes.iter() {
-        for reduction_count in reduction_counts.iter() {
+    for reduction_count in reduction_counts.iter() {
+        let mut group: BenchmarkGroup<'_, _> =
+            c.benchmark_group(format!("LEM Fibonacci Prove - rc = {}", reduction_count));
+        group.sampling_mode(SamplingMode::Flat); // This can take a *while*
+        group.sample_size(10);
+        group.noise_threshold(noise_threshold_env().unwrap_or(0.05));
+
+        for fib_n in batch_sizes.iter() {
             let prove_params = ProveParams {
                 fib_n: *fib_n,
                 reduction_count: *reduction_count,
                 date: env!("VERGEN_GIT_COMMIT_DATE"),
                 sha: env!("VERGEN_GIT_SHA"),
             };
-            fibo_prove(prove_params, &mut group, &state);
+            fibonacci_prove(prove_params, &mut group, &state);
         }
     }
 }
@@ -161,7 +196,7 @@ cfg_if::cfg_if! {
             .sample_size(10)
             .with_profiler(pprof::criterion::PProfProfiler::new(100, pprof::criterion::Output::Flamegraph(None)));
             targets =
-             fibonacci_prove,
+             fibonacci_benchmark,
          }
     } else {
         criterion_group! {
@@ -170,7 +205,7 @@ cfg_if::cfg_if! {
             .measurement_time(Duration::from_secs(120))
             .sample_size(10);
             targets =
-             fibonacci_prove,
+             fibonacci_benchmark,
          }
     }
 }
